Se agregaron métodos y atributos para acceder a variables difusas
El objeto FIS (y FuzzyRule) contiene atributos para obtener las entradas (antecedentes) y salidas (consecuentes) una vez que es instanciado; la estructura generada es un diccionario.

Las clases Antecedent y Agregation contienen el atributo fuzzy_variables que regresan, en forma de diccionario, las variables difusas que contienen

+++ FuzzySystem/matlab_importer.py
@@ -93,7 +93,8 @@ def read_fis_file(file):
                     key, value = text.strip().split('=')
                     if key=='Range':
                         value = value.replace('[','').replace(']','')
-                        value = list(map(int,value.split(' ')))
+                        #value = list(map(int,value.split(' ')))
+                        value = list(map(float,value.split(' ')))
                         temp[key] = value
                     elif key == 'NumMFs':
                         temp[key] = int(value)
